Move Report interface to types folder
