+++ ui/src/SortingAndFiltering/FilterLibraries.ts
@@ -39,7 +39,6 @@ export interface FilterType {
     tagNameType: TagNameType;
 }
 
-export type LabelType = 'Location Tags:' | 'Product Tags:' | 'Role Tags:' | 'Person Tags:';
 export type TagType = 'role' | 'product tag' | 'location' | 'person tag';
 export type TagNameType = 'Role' | 'Product Tag' | 'Location' | 'Person Tag';
